ARROW-1618: [JAVA] Reduce Heap Usage (Phase 1)
cc @jacques-n , @icexelloss , @BryanCutler

This is initial small phase of our attempt to reduce heap usage per vector.

As part of investigation, we realized its better to address few things as part of subtasks for ARROW-1463. Accordingly, I need to update the requirements document w.r.t heap usage for ARROW-1471.

This patch gets rid of Release Listener object in Allocation Manager as all the logic is implemented as part of AllocationManager itself.

https://docs.google.com/document/d/1MU-ah_bBHIxXNrd7SkwewGCOOexkXJ7cgKaCis5f-PI/edit

Author: siddharth <siddharth@dremio.com>

Closes #1142 from siddharthteotia/ARROW-1618 and squashes the following commits:

77151a27 [siddharth] ARROW-1618: Reduce Heap Usage (Phase 1)