ARROW-1047: [Java] Add Generic Reader Interface for Stream Format
This change decouples the reading of messages from the ReadChannel so that it is possible to build a reader that is not tied to a specific stream.  This adds a new interface `MessageReader` that will return a `Message` and the message body.  The `MessageChannelReader` implements this interface to read from a `ReadChannel` to match the current functionality.  A re-org of reading and writing packages is also done to better organize classes under an `ipc` package.

There is a slight change in behavior for the `ArrowFileReader` that should not affect any usage.  Previously,  the schema was read during initialization, and all dictionaries read just before the first record batch was read.  Now, all dictionaries are read directly after the schema, and not specifically tied to reading the first record batch.

Author: Bryan Cutler <cutlerb@gmail.com>

Closes #1259 from BryanCutler/java-generic-stream-interfaces-ARROW-1047 and squashes the following commits:

43314ca1 [Bryan Cutler] ARROW-1047: [Java] Add Generic Reader Interface for Stream Format