ARROW-5136: [Flight] Call options
Right now the only option is a timeout on an overall call, which will be useful to limit hangs when connecting to a Flight service that might have gone down. (gRPC is rather aggressive about retrying for a while before giving up.)

You could imagine having a health check action in DoAction, for instance, which fails if the call doesn't complete within a short time.

I would also like a stream timeout, where the timeout only fires if there is a long delay between messages on a DoGet or DoAction, but this is hard to implement as-is in gRPC. Implementing an async client interface might be a better way to achieve this.

Author: David Li <li.davidm96@gmail.com>

Closes #4144 from lihalite/arrow-5136-call-options and squashes the following commits:

f20181c90 <David Li> Use varargs for Java Flight CallOptions
a573d2c61 <David Li> Make FlightCallOptions a POD type
d4bef2781 <David Li> Add tests for Flight call options in Python
0e6111154 <David Li> Don't hold GIL during Flight server shutdown
a1829b32c <David Li>  implement call options for Flight
81e3c7908 <David Li>  add call options to Flight