ARROW-5583: [Java] When the isSet of a NullableValueHolder is 0, the buffer field should not be used
For each variable-width vector, like the VarCharVector, it has a set method that uses a NullableValueHolder as the input parameter. When the isSet field is set to 0, it means the value to set is null, so the buffer field of the NullableValueHolder is invalid, and should not be used.

For example, the user may set a null value in the VarCharVector with the following code snippet:

NullableVarCharHolder holder = new NullableVarCharHolder();
holder.isSet = 0;
...
varCharVector.set(i, holder);

Please note that in the code above, the holder.buffer is not set, so it is null. According to the VarCharVector#set method, it will set the bytes using holder.buffer even if holder.isSet equals 0. This will lead to an exception.

Author: liyafan82 <fan_li_ya@foxmail.com>

Closes #4543 from liyafan82/fly_0613_varlen and squashes the following commits:

d792e0a38 <liyafan82>  Revise the code for setSafe methods
43a51c20b <liyafan82>  Resolve comments
40c1548e6 <liyafan82>  When the isSet of a NullableValueHolder is 0, the buffer field should not be used