ARROW-1279: [Integration] Enable MapType integration tests
This PR enables MapType integration tests for C++ and Java, and fixes remaining issues to pass. Previously, C++ created a MapType with a StructType field that is nullable, which is required to be non-nullable from the spec. The struct field has the name "entries" and the struct data pairs are named "key" and "value."

Author: Bryan Cutler <cutlerb@gmail.com>

Closes #4684 from BryanCutler/map-type-integration-tests-ARROW-1279 and squashes the following commits:

6259a4574 <Bryan Cutler> rename MapType fields for internal test
30777804f <Bryan Cutler> rename MapType struct field to 'entries'
3185af9fa <Bryan Cutler> Fix MapType to create a non-nullable struct field