ARROW-6070: [Java] Avoid creating new schema before IPC sending
If a dictionary is attached to a schema, it may need to be converted before IPC sending. When this is not the case (which is most likely in practice), there is no need to do the conversion and no need to create a new schema.

We solve the above problem by quickly determining if conversion is required, and if not, we avoid creating a new schema and return the original one immediately.

Closes #4968 from liyafan82/fly_0730_schema and squashes the following commits:

514955bd7 <liyafan82>  Avoid creating new schema before IPC sending

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>