ARROW-7429: [Java] Enhance code style checking for Java code (remove consecutive spaces)
This issue is opened in response to a discussion in https://github.com/apache/arrow/pull/5861#discussion_r348917065.

We found the current style checking for Java code is not sufficient. So we want to enhace it in a series of "small" steps, in order to avoid having to change too many files at once.

In this issue, we remove consecutive spaces between tokens, so that tokens are separated by single spaces.

Closes #6060 from liyafan82/fly_1217_style and squashes the following commits:

d14b8d572 <emkornfield> Merge branch 'master' into fly_1217_style
6d4fe3d77 <liyafan82>  Enhance code style checking for Java code (remove consecutive spaces)

Lead-authored-by: liyafan82 <fan_li_ya@foxmail.com>
Co-authored-by: emkornfield <emkornfield@gmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>