ARROW-17055: [Java][FlightRPC] Don't duplicate generated Protobuf classes between flight-core and flight-sql (#13596)
flight-core and flight-sql jars delivering same class names. 

It seems that the classes generated by Flight.proto gets generated in both flight-sql and flight-core jars. Since these classes get generated in flight-core, and flight-sql is dependent on flight-core, can the generation of Flight.java and FlightServiceGrpc.java be removed from flight-sql and instead rely on it to be pulled directly from flight-core.

Authored-by: KBambrick1 <kevinbambrick1@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>