GH-33993: [Java] Let OS assign port in tests while creating Flight server (#33992)
### Rationale for this change

Sometimes tests are flaky because the current port selection algorithm selects a random port between the ranges 49152-54152. This change makes it such that the OS will assign the next free port without us searching for a port in this range.

### What changes are included in this PR?

This PR makes use of  `java.net.ServerSocket`'s ability to assign the next free port if 0 is passed as the port number.

### Are these changes tested?

If existing tests pass, there's confirmation that the Flight server is able to come up successfully, thereby no new tests should be needed.

### Are there any user-facing changes?

No, there are no user facing changes.

* Closes: #32954
* Closes: #33993

Authored-by: Ramasai <ramasai.tadepalli+3108@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>