GH-35275: [Java] Ensure VectorSchemaRoot slice returns a new root (#35476)
### Rationale for this change
The optimization in the `slice` function for VectorSchemaRoot that returns the reference to the original root when the slice covers the entire root should be removed because this is an inconsistent interface (sometimes you get a totally new autocloseable and sometimes you get the original one). 

Also - in the [this documentation page](https://arrow.apache.org/docs/java/vector_schema_root.html#vectorschemaroot) it explicitly says that "A **NEW** VectorSchemaRoot can be sliced from an existing root without copying data:" so any time you call `slice`, the slice root should be a new one (even if it's the entire root). 

The test for slice has also been changed to have more coverage of the slice function (not just slicing from the start index).  

### What changes are included in this PR?
Remove self reference return in `VectorSchemaRoot#split` and expand test coverage for the function.

### Are these changes tested?
Yes

### Are there any user-facing changes?
Yes 

**This PR includes breaking changes to public APIs.** 

* Closes: #35275

Authored-by: Parth <parthchonkar@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>