GH-37704: [Java] Add schema IPC serialization methods (#37778)
### Rationale for this change

The methods in the Schema class in Java serialize the schema in a way that is inconsistent with other languages. This can cause integration issues when using a Java-serialized schema in another language via IPC or Arrow Flight or vice-versa.

### What changes are included in this PR?

Added the serializeAsMessage() and deserializeMessage() methods to Schema for serialization Schemas to the standard IPC format (wrapped in an IPC Message).

Marked the methods that serialize to raw FlatBuffer objects as deprecated (toByteArray() and deserialize()).

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.
* Closes: #37704

Authored-by: James Duong <duong.james@gmail.com>
Signed-off-by: David Li <li.davidm96@gmail.com>