MINOR: Add explicit exception when no more buffer can be read when loading buffers (#649)
## What's Changed

When `VectorLoader` tries to load buffers (i.e., `loadBuffers`), it has
detect on the error case that some buffers are not consumed. But another
error that the number of buffers is less than expected is not handled
for now. Once it is happened, users will get
`java.util.NoSuchElementException` which is not easy to understand.

This patch adds an explicit exception for such case.

See more discussion at #648.