GH-655: Failure in UnionReader.read after DecimalVector promotion to UnionVector (#656)
When a DecimalVector is promoted to a UnionVector via a
PromotableWriter, the UnionVector will have the decimal vector in it's
internal struct vector, but the decimalVector field will not be set. If
UnionReader.read is then used to read from the UnionVector, it will fail
when it tries to read one of the promoted decimal values, due to
decimalVector being null, and the exact decimal type not being provided.
This failure is unnecessary though as we have a pre-existing decimal
vector, the caller just does not know the exact type - and it shouldn't
be required to.

The change here is to check for a pre-existing decimal vector in the
internal struct when getDecimalVector() is called. If one exists, set
the decimalVector field and return. Otherwise, if none exists, throw the
exception.

## What's Changed

Updated UnionVector to handle this case.

Closes GH-655

---------

Co-authored-by: Scott Cowell <scott.cowell@dremio.com>