GH-661: [Flight] JDBC: Cache failed locations (#662)
## What's Changed

Instead of repeatedly trying to access locations which have failed
before, cache them and move them to the end of the list of locations to
attempt on subsequent accesses. This provides a better user experience
for Flight SQL services that return multiple locations in the
expectation that the user may not be able to access some of them (but
would have a better experience if they are).

Requires #494.

Closes #661.